pyproject.toml
[project]
name = "Textbox"
version = "0.1.0"
description = "Render your texts into textbot and send."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "keyboard>=0.13.5",
    "pillow>=12.0.0",
    "pip>=25.3",
    "psutil>=7.1.3",
    "pyperclip>=1.11.0",
    "pywin32>=311",
    "setuptools>=80.9.0",
]
